bugs fixed

+++ code_producers/src/c_elements/common64/circom.hpp
@@ -9,7 +9,7 @@
 
 //#include "fr.hpp"
 
-typedef unsigned long long u64;
+typedef uint64_t u64;
 typedef uint32_t u32;
 typedef uint8_t u8;
